main: move libevent api callbacks to separate file
The main.c file contains both command line parsing, the main() function
*and* libevent stuff for managing the API socket. This is too many
purposes for a single file.

Move the API socket handling in a separate file. Adjust stuff
accordingly.
